#82 consolid the common fields within symbol def and unit def into a common base class, which is useful for building a symbol picker
